MDL-71242 core_course: Validate the value of the sort argument
This change validates the value of the sort argument in
course_get_recent_courses(). As this value is directly concatenated
onto the constucted SQL query string, a prior validation is required
in order to avoid a potential SQL injection or passing incorrect
values for the sorting columns or sorting directions.